requirements.txt
finnhub-python
pandas
requests
python-dotenv
jupyter
ipykernel
matplotlib
websockets
asyncio